backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Jinja2
MarkupSafe
SQLAlchemy
Werkzeug
astroid
click
dill
ecdsa
future
greenlet
isort
itsdangerous
jose
lazy-object-proxy
mccabe
pip
platformdirs
pycodestyle
pycryptodome
pylint
python-dotenv
python-jose
python-jose-cryptodome
setuptools
six
tomlkit
typed-ast
typing-extensions
wheel
wrapt